Fix sockets being recreated on every transition.  Will fix #602.
