+++ src/containers/App/index.tsx
@@ -11,18 +11,54 @@ import AppErrorBoundary from './AppErrorBoundary'
 import NoMatch from '../NoMatch'
 import { Header } from '../Header'
 import { queryClient } from '../shared/QueryClient'
+import { RouteDefinition } from '../shared/routing'
+import {
+  ACCOUNT,
+  LEDGER,
+  LEDGERS,
+  NETWORK,
+  NFT as NFTRoute,
+  PAYSTRING,
+  TOKEN,
+  TRANSACTION,
+  VALIDATOR,
+} from './routes'
+import Ledgers from '../Ledgers'
+import { Ledger } from '../Ledger'
+import { AccountsRouter } from '../Accounts/AccountsRouter'
+import { Transaction } from '../Transactions'
+import { Network } from '../Network'
+import { Validator } from '../Validators'
+import { PayString } from '../PayStrings'
+import Token from '../Token'
+import { NFT } from '../NFT/NFT'
+import { LegacyRoutes } from './LegacyRoutes'
 
 export const AppWrapper = () => {
   const { t } = useTranslation()
   const location = useLocation()
   const mode = process.env.VITE_ENVIRONMENT
 
-  const rippledUrl =
-    process.env.VITE_ENVIRONMENT === 'custom'
-      ? `/${location.pathname.split('/')[1]}`
-      : ''
+  const rippledUrl = mode === 'custom' ? location.pathname.split('/')[1] : ''
+  const basename = mode === 'custom' ? `/${rippledUrl}` : ''
+  const updatePath = (path) => `${basename}${path}`
+
+  // Defined here rather than ./routes to avoid circular dependencies when using RouteDefinitions with <RouteLink>.
+  const routes: [RouteDefinition<any>, any][] = [
+    [LEDGERS, Ledgers],
+    [LEDGER, Ledger],
+    [ACCOUNT, AccountsRouter],
+    [TRANSACTION, Transaction],
+    [NETWORK, Network],
+    [VALIDATOR, Validator],
+    [PAYSTRING, PayString],
+    [TOKEN, Token],
+    [NFTRoute, NFT],
+  ]
+
   return (
     <HelmetProvider>
+      <LegacyRoutes basename={basename} />
       <QueryClientProvider client={queryClient}>
         <div className="app-wrapper">
           <AppErrorBoundary>
@@ -39,8 +75,16 @@ export const AppWrapper = () => {
                 {mode === 'custom' && (
                   <Route path="/" element={<CustomNetworkHome />} />
                 )}
-                <Route path="/*" element={<App />} />
-                <Route element={<NoMatch />} />
+                <Route element={<App rippledUrl={rippledUrl} />}>
+                  {routes.map(([route, Component]) => (
+                    <Route
+                      key={route.path}
+                      path={updatePath(route.path)}
+                      element={<Component />}
+                    />
+                  ))}
+                  <Route path="*" element={<NoMatch />} />
+                </Route>
               </Routes>
               <Footer />
             </div>
